lib(photobooth): move Photobooth class out of helper
