+++ welcome.php
@@ -1,6 +1,6 @@
 <?php
 require_once 'lib/config.php';
-require_once 'lib/helper.php';
+require_once 'lib/photobooth.php';
 
 if (!is_file('.skip_welcome')) {
     touch('.skip_welcome');
